authentication, hide buttons for unauthenticated user, reset to cero half way

+++ client/components/Carrusel.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
 import Picture from './Picture';
+import { connect } from 'react-redux';
+
 
 class Carrusel extends Component {
 
@@ -9,12 +11,12 @@ class Carrusel extends Component {
     this.state = {
       show: true,
       frontState: 'show',
-      currentSlide: 0,
       data: this.props.details,
+      currentSlide: 0
     }
 
-
     this.toggleCard = this.toggleCard.bind(this);
+    this.showItem = this.showItem.bind(this);
 
     this.frontOptions = this.frontOptions.bind(this);
 
@@ -32,15 +34,45 @@ class Carrusel extends Component {
     this.toggleNext = this.toggleNext.bind(this);
     this.togglePrev = this.togglePrev.bind(this);
 
+    this.canSee = this.canSee.bind(this);
+    this.menuButtons = this.menuButtons.bind(this);
+    this.makeTheCero = this.makeTheCero.bind(this);
+
   }
+
+  canSee() {
+    // console.log('can see')
+    if(this.props.user.role == 'admin')
+      return true
+  }
+
+  menuButtons() {
+    // console.log('menu buttons')
+    if(this.canSee()) {
+      return(
+        <div>
+          <i type="button" onClick={this.addItem} className="pic-options small material-icons">add</i>
+          <i type="button" onClick={this.editItem} className="pic-options small material-icons">mode_edit</i>
+          <i type="button" onClick={this.deleteItem} className="pic-options small material-icons">delete</i>
+        </div>
+      )
+    }
+  }
+
+  makeTheCero() {
+    this.setState({currentSlide: 0 })
+  }
+
 
   toggleCard() {
     this.setState({ show: !this.state.show });
   }
 
+  // console.log(this.props.reset)
+
   componentDidMount(){
     // Text of each picture
-    $('.carousel-item').on('mouseenter', function() {
+    $('.carruslide').on('mouseenter', function() {
       var info = this.dataset.info || "";
       if(info != "") {
         $('.letter').text(info);
@@ -48,14 +80,22 @@ class Carrusel extends Component {
       };
     });
 
-    $('.carousel').on('mouseleave', function() {
+    $('.carruslide').on('mouseleave', function() {
       $('.letter').hide();
     });
   }
 
   componentDidUpdate(){
+    // reset current slide to cero!
+    if(this.props.reset == true) {
+      console.log('here')
+      this.setState({
+        currentSlide: 0
+      })
+      this.props.reseter();
+    }
     // Text of each picture
-    $('.carousel-item').on('mouseenter', function() {
+    $('.carruslide').on('mouseenter', function() {
       var info = this.dataset.info || "";
       if(info != "") {
         $('.letter').text(info);
@@ -63,24 +103,28 @@ class Carrusel extends Component {
       };
     });
 
-    $('.carousel').on('mouseleave', function() {
+    $('.carruslide').on('mouseleave', function() {
       $('.letter').hide();
     });
   }
 
+  componentWillMount() {
+
+  }
+
   showMeThePic() {
+    console.log(this.state.currentSlide)
     const info = this.props.details
     const actualPic = this.state.currentSlide;
     const id = info[actualPic].id;
 
-    return (<img className="carousel-item" src={info[actualPic].image} data-info={info[actualPic].infopic} id={id} />)
+    return (<img className="carruslide" src={info[actualPic].image} data-info={info[actualPic].infopic} id={id} />)
   }
 
   toggleNext() {
     // console.log("this is toggle next");
     // console.log(this.state.data);
     const data = this.state.data;
-
     const current = this.state.currentSlide;
     // console.log(current)
     let next = current + 1;
@@ -88,7 +132,8 @@ class Carrusel extends Component {
       next = 0;
     }
     console.log(next);
-    this.setState({ currentSlide: next });
+    this.setState({currentSlide: next});
+    this.props.reseter()
     // console.log(this.state.currentSlide)
   }
 
@@ -102,7 +147,8 @@ class Carrusel extends Component {
       prev = data.length - 1;
     }
     console.log(prev);
-    this.setState({ currentSlide: prev });
+    this.setState({currentSlide: prev});
+    this.props.reseter()
     // console.log(this.state.currentSlide)
   }
 
@@ -114,7 +160,7 @@ class Carrusel extends Component {
     return (
       <div>
         <form ref="imageForm" className="input-field">
-          <h4 className="center title">Add item</h4>
+          <span className="center title"><i type="button" onClick={this.showItem} className="close material-icons right">close</i><h5>Add new image</h5></span>
           <div className="card-content">
             <span className="card-title">
               <input type="text" required ref='name'  placeholder="name" />
@@ -165,7 +211,7 @@ class Carrusel extends Component {
     return (
       <div>
         <form ref="imageForm" className="input-field">
-          <h4 className="center title">Add item</h4>
+          <span className="center title"><i type="button" onClick={this.showItem} className="close material-icons right">close</i><h5>Edit {image.name}</h5></span>
           <div className="card-content">
             <span className="card-title">
               <input type="text" required ref='name'  placeholder="name" defaultValue={image.name}/>
@@ -206,10 +252,12 @@ class Carrusel extends Component {
         console.log(data)
     })
   }
-  
+
   deleteItem(e) {
     e.preventDefault();
     let id = this.props.details[this.state.currentSlide].id
+    let yes = confirm('are you sure');
+    if (yes == true) {
       $.ajax({
         url: `/api/carrusels/${id}`,
         type: 'DELETE',
@@ -220,44 +268,47 @@ class Carrusel extends Component {
         })
       }).fail( data => {
         console.log(data)
-    })
+      });
+    }
   }
 
   frontOptions() {
+    const actualPic = this.state.currentSlide;
     if(this.state.frontState === 'show') {
       return(
-        <div className="card-image waves-effect waves-block waves-light">
-          <div className="prev detalles valign-wrapper" onClick={this.togglePrev}></div>
-            <div id="detalles" className="carousel carousel-slider">
+        <div >
+          <div>
+            <div id="detalles" className="carousel carruholder center">
+              <div className="prev detalles valign-wrapper" onClick={this.togglePrev}></div>
               {this.showMeThePic(this.props.details)}
+              <div className="letter"></div>
+              <div className="next detalles valign-wrapper" onClick={this.toggleNext}></div>
             </div>
-          <div className="letter"></div>
-          <div className="next detalles valign-wrapper" onClick={this.toggleNext}></div>
+          </div>
+          <div className="card-info">
+            <h5 className="card-title left">{this.props.details[actualPic].name}</h5>
+            <span className="rigth">
+              {this.menuButtons()}
+              <i type="button" onClick={this.toggleCard} className="hamburger right"></i>
+            </span>
+          </div>
         </div>
       );
     } else if (this.state.frontState === 'add') {
-      return(this.addFrom());
+      return(this.addForm());
     } else {
       return(this.editForm())
     }
   }
 
+  showItem() {
+    this.setState({ frontState: 'show' });
+  }
 
   front() {
-    const actualPic = this.state.currentSlide;
     return(
-      <div>
-        <div className="card">
-          { this.frontOptions()}
-          <div className="card-content">
-            <span className="card-title activator grey-text text-darken-4">{this.props.details[actualPic].name}
-              <div type="button" onClick={this.toggleCard} className="hamburger right"></div>
-              <div type="button" onClick={this.addItem}><i className="small material-icons">add</i></div>
-              <div type="button" onClick={this.editItem}><i className="small material-icons">mode_edit</i></div>
-              <div type="button" onClick={this.deleteItem}><i className="small material-icons">delete</i></div>
-            </span>
-          </div>
-        </div>
+      <div className="card">
+        { this.frontOptions()}
       </div>
     )
   }
@@ -266,7 +317,7 @@ class Carrusel extends Component {
     return(
       <div>
         <div className="card-reveal">
-          <span className="card-title grey-text text-darken-4">{this.props.details.name}<i type="button" onClick={this.toggleCard} className="material-icons right">close</i></span>
+          <span className="card-title grey-text text-darken-4"><i type="button" onClick={this.toggleCard} className="close material-icons right">close</i><h5>{this.props.description.name}</h5></span>
           <p>{this.props.description.description}</p>
           <div className="logos">
             <div className={this.props.description.logo1}></div>
@@ -291,4 +342,10 @@ class Carrusel extends Component {
 }
 
 
-export default Carrusel;
+const mapStateToProps = (state) => {
+  return {
+    user: state.user
+  }
+}
+
+export default connect(mapStateToProps)(Carrusel);
